+++ R/app_server.R
@@ -50,4 +50,7 @@ app_server <- function(input, output, session) {
       pop = 428826
     )
     mod_focus_20200323_picco_server("picco")
+    mod_focus_20200325_hosp_server("hosp")
+    mod_focus_20200328_hosp_server("tot")
+    mod_focus_20200331_server("ven_pie")
 }