+++ R/eng-app_server.R
@@ -49,4 +49,8 @@ eng_app_server <- function(input, output, session) {
                                      loc = "Alessandria",
                                      pop = 428826
     )
+    eng_mod_focus_20200323_picco_server("picco")
+    eng_mod_focus_20200325_hosp_server("hosp")
+    eng_mod_focus_20200328_hosp_server("tot")
+    eng_mod_focus_20200331_server("ven_pie")
 }