+++ R/eng-app_ui.R
@@ -35,24 +35,14 @@ eng_dashboard_header <- function() {dashboardHeader(
   title = "COVID-19 - Italy",
 
   dropdownMenu(type = "messages",
-    messageItem(
-      from = "Highlights",
-      message = "2020-03-21 Alessandria: cases predictions",
-      icon = icon("search")
-    ),
-    messageItem(
-      from = "Highlights",
-      message = "2020-03-21 Vercelli: cases predictions",
-      icon = icon("search")
-    ),
-    messageItem(
-      from = "Highlights",
-      message = "2020-03-21 Novara: cases predictions",
-      icon = icon("search")
-    ),
-    messageItem(
+messageItem(
+  from = "Highlights",
+  message = "2020-03-31 Comparative Ven-Pie",
+  icon = icon("search")
+  ),
+messageItem(
       from = "Data",
-      message = "Data updated (2020-03-19 18:20)",
+      message = glue::glue("Data updated ({last_data_update})"),
       icon = icon("database")
     ),
     messageItem(
@@ -82,6 +72,18 @@ eng_dashboard_sidebar <- function() {dashboardSidebar(sidebarMenu(
   menuItem("Home", tabName = "home", icon = icon("home")),
 
   menuItem("Highlights", icon = icon("bullseye"),
+           menuSubItem("2020-03-31 Comparative", tabName = "20200331Comp",
+                       icon = icon("flag")
+           ),
+           menuSubItem("2020-03-28 Veneto", tabName = "20200328Tot_hosp",
+                       icon = icon("flag")
+           ),
+           menuSubItem("2020-03-25 Veneto", tabName = "20200325Hosp",
+                       icon = icon("flag")
+           ),
+           menuSubItem("2020-03-24 Methodology", tabName = "20200323Picco",
+                       icon = icon("flag")
+           ),
            menuSubItem("2020-03-21 Alessandria", tabName = "20200321Alessandria",
                        icon = icon("flag")
            ),
@@ -168,7 +170,7 @@ eng_dashboard_body <- function() {dashboardBody(
             of Cardiac, Thoracic and Vascular Sciences and Public Health --
             Università degli Studi di Padova."
           ),
-          a(href = "https://linkedin.com/in/dario-gregori-2720039", target = "_blank", "LinkedIn")
+          a(href = "https://linkedin.com/in/dario-gregori-2720039/", target = "_blank", "LinkedIn")
         )
       ),
 
@@ -177,22 +179,22 @@ eng_dashboard_body <- function() {dashboardBody(
           Biostatistics, Epidemiology, and Public Health of the Department
           of Cardiac, Thoracic and Vascular Sciences and Public Health --
           Università degli Studi di Padova. Head of the Laboratory of Artificial Intelligence for Medical Sciences"),
-          a(href = "https://linkedin.com/in/corradolanera", target = "_blank", "LinkedIn")
+          a(href = "https://linkedin.com/in/corradolanera/", target = "_blank", "LinkedIn")
         )
       ),
 
 
-      box(width = 12, title = HTML("<strong>Epidemiological modelling</strong>"),
+      box(width = 12, title = HTML("<strong>Epidemiological modeling</strong>"),
         p(
           HTML("Prof. <strong>Paola Berchialla</strong>, PhD, Department of Clinical
             and Biological Sciences -- Università degli Studi di Torino"
           ),
-          a(href = "https://linkedin.com/in/paola-berchialla-36b44410", target = "_blank", "LinkedIn")
+          a(href = "https://linkedin.com/in/paola-berchialla-36b44410/", target = "_blank", "LinkedIn")
         ),
 
         p(
           HTML(
-            '<strong>Dolores Catelan</strong>, Ph.D., Department of
+            'Prof. <strong>Dolores Catelan</strong>, Ph.D., Department of
             Statistics, Computer Sciences, Applications "G. Parenti" (DISIA),
             Università degli Studi di Firenze'
           ),
@@ -203,35 +205,50 @@ eng_dashboard_body <- function() {dashboardBody(
         )
       ),
 
-      box(width = 12, title = HTML("<strong>Modelli Previsivi</strong>"),
+      box(width = 12, title = HTML("<strong>Predictive models</strong>"),
         p(HTML("<strong>Danila Azzolina</strong>, PhD, Department of Translational Medicine --
           Università del Piemonte Orientale"),
-          a(href = "https://linkedin.com/in/danila-azzolina-862465166", target = "_blank", "LinkedIn")
+          a(href = "https://linkedin.com/in/danila-azzolina-862465166/", target = "_blank", "LinkedIn")
         ),
 
         p(HTML("<strong>Ilaria Prosepe</strong>, MSc, Unit of
            Biostatistics, Epidemiology, and Public Health of the Department
            of Cardiac, Thoracic and Vascular Sciences and Public Health --
            Università degli Studi di Padova."),
-          a(href = "https://linkedin.com/in/ilaria-prosepe-1b52371a4", target = "_blank", "LinkedIn")
+          a(href = "https://linkedin.com/in/ilaria-prosepe-1b52371a4/", target = "_blank", "LinkedIn")
         )
       ),
 
 
-      box(width = 12, title = HTML("<strong>Modelli di Environmental Epidemiology and Pollution</strong>"),
+      box(width = 12, title = HTML("<strong>Modeling in Environmental Epidemiology and Pollution</strong>"),
           p(
             HTML(
               'Prof. <strong>Annibale Biggeri</strong>, MD, MSPH, MSB,
               Department of Statistics, Computer Sciences, Applications
               "G. Parenti" (DISIA), Università degli Studi di Firenze'
             ),
-            a(href = "https://linkedin.com/in/danila-azzolina-862465166", target = "_blank", "LinkedIn")
+            a(href = "https://linkedin.com/in/danila-azzolina-862465166/", target = "_blank", "LinkedIn")
           ),
-          p(HTML("<strong>Elisa Gallo</strong>, MS., Unità di
+          p(
+            HTML(
+              'Prof. <strong>Cristina Canova</strong>, Ph.D., Unit of
+           Biostatistics, Epidemiology and Public Health of the Department
+           of Cardiac, Thoracic and Vascular Sciences and Public Health --
+           Università degli studi di Padova.'
+            ),
+            a(href = "https://www.linkedin.com/in/cristina-canova-05448861/", target = "_blank", "LinkedIn")
+          ),
+          p(HTML("<strong>Elisa Gallo</strong>, MS., Unit of
            Biostatistics, Epidemiology and Public Health of the Department
            of Cardiac, Thoracic and Vascular Sciences and Public Health --
            Università degli studi di Padova."),
             a(href = "https://www.linkedin.com/in/elisa-gallo-9b3933152/", target = "_blank", "LinkedIn")
+          ),
+          p(HTML("<strong>Francesco Garzotto</strong>, MSc, Unit of
+           Biostatistics, Epidemiology and Public Health of the Department
+           of Cardiac, Thoracic and Vascular Sciences and Public Health --
+           Università degli studi di Padova."),
+            a(href = "https://www.linkedin.com/in/francesco-garzotto-19907826/", target = "_blank", "LinkedIn")
           )
       ),
 
@@ -243,7 +260,7 @@ eng_dashboard_body <- function() {dashboardBody(
             of Cardiac, Thoracic and Vascular Sciences and Public Health --
             Università degli Studi di Padova. Head of the Laboratory of Clinical Epidemiology and Digital Health"
           ),
-          a(href = "https://linkedin.com/in/giulia-lorenzoni-b382a6180", target = "_blank", "LinkedIn")
+          a(href = "https://linkedin.com/in/giulia-lorenzoni-b382a6180/", target = "_blank", "LinkedIn")
         ),
         p(
           HTML(
@@ -281,7 +298,22 @@ eng_dashboard_body <- function() {dashboardBody(
     )
 
   ),
-
+  tabItem(tabName = "20200331Comp",
+          h2("Comparative analysis between the Piemonte Region and Veneto Region of the epidemiological data relative to Covid-19 infection."),
+          eng_mod_focus_20200331_ui("ven_pie")
+  ),
+  tabItem(tabName = "20200328Tot_hosp",
+          h2("Possible effect on hospitalizations of the health policies implemented by the Veneto region"),
+          eng_mod_focus_20200328_hosp_ui("tot")
+  ),
+  tabItem(tabName = "20200325Hosp",
+          h2("Possible effect on hospitalizations of the health policies implemented by the Veneto region"),
+          eng_mod_focus_20200325_hosp_ui("hosp")
+  ),
+  tabItem(tabName = "20200323Picco",
+          h2("Impact of statistical uncertainty on COVID-19 predictions"),
+          eng_mod_focus_20200323_picco_ui("picco")
+  ),
   tabItem(tabName = "20200321Alessandria",
           h2("Expected number of total cases in Alessandria"),
           eng_mod_focus_20200320_novara_ui("da_alessandria")
@@ -297,56 +329,67 @@ eng_dashboard_body <- function() {dashboardBody(
   ),
 
   tabItem(tabName = "20200319Veneto",
-          h2("Possible impact of the health policies implemented in the Veneto region on the number of patients admitted to the ICU"),
+          h2("Possible effect on ICU admissions of the health policies implemented by the Veneto region"),
           eng_mod_focus_20200318_veneto_intensive_ui("21")
   ),
 
   tabItem(tabName = "20200318Piemonte",
-          h2("Possible effect of the health policies implemented in the Piemonte region"),
+          h2("Possible effect of the health policies implemented by the Piemonte region"),
           eng_mod_focus_20200318_piemonte_ui("20200318_piemonte")
   ),
   tabItem(tabName = "20200318Fvg",
-          h2("Possible effect of the health policies implemented in the Friuli Venezia Giulia region"),
+          h2("Possible effect of the health policies implemented by the Friuli Venezia Giulia region"),
           eng_mod_focus_20200318_friuli_ui("20200318_fvg")
   ),
   tabItem(tabName = "20200314Veneto",
-    h1("Possible effect of the health policies implemented in the Veneto region"),
+    h1("Possible effect of the health policies implemented by the Veneto region"),
     eng_mod_focus_20200314_ui("dapb")
   ),
 
   tabItem(tabName = "national",
     h2("National events"),
-    mod_ts_ita_ui("ts_nat_cum", title = "Time series trend - cumulative events"),
-    mod_ts_ita_ui("ts_nat_inc", title = "Time series trend - daily new events")
+    box(width = 12, title = "Instructions",
+        p("Visualize/hide one or more measures on the graph by clicking on the items in the legend. Double click to only visualize the selected item."),
+        p("Click on the autoscale button (the third one) to maximize the size of the graph."),
+    ),
+    eng_mod_ts_ita_ui("ts_nat_cum", title = "Time series trend - cumulative events"),
+    eng_mod_ts_ita_ui("ts_nat_inc", title = "Time series trend - daily new events")
   ),
 
 
   tabItem(tabName = "regional",
     h2("Regional events"),
+    box(width = 12, title = "Istructions",
+        p("Add/remove computations for one or more region/measure by adding/removing it from the box."),
+        p("NOTE: The number of tests per day can be visualized on the graph (by selecting the correct item in the legend) but it is hidden by default as it is off the scale (compared to the other measures)."),
+        p(""),
+        p("Visualize/hide one or more regions/measures on the graph by clicking on the items in the legend. Double click to only visualize the selected item."),
+        p("Click on the autoscale button (the third one) to maximize the size of the graph."),
+      ),
 
     h3("Time series by region"),
     fluidRow(
       box(title = "Cumulative events", width = 12,
-        mod_ts_reg_ui("ts_reg_cum_mes")
+        eng_mod_ts_reg_ui("ts_reg_cum_mes")
       )
     ),
 
     fluidRow(
       box(title = "Daily new events", width = 12,
-        mod_ts_reg_ui("ts_reg_inc_mes")
+        eng_mod_ts_reg_ui("ts_reg_inc_mes")
       )
     ),
 
     h3("Regional time series by event"),
     fluidRow(
       box(title = "Cumulative events", width = 12,
-        mod_ts_reg_ui("ts_reg_cum_reg")
+        eng_mod_ts_reg_ui("ts_reg_cum_reg")
       )
     ),
 
     fluidRow(
       box(title = "Daily new events", width = 12,
-        mod_ts_reg_ui("ts_reg_inc_reg")
+        eng_mod_ts_reg_ui("ts_reg_inc_reg")
       )
     )
 
@@ -355,17 +398,20 @@ eng_dashboard_body <- function() {dashboardBody(
 
   tabItem(tabName = "provincial",
     h2("Provincial events"),
+    p("Add/remove computations for one or more region/measure by adding/removing it from the box."),
+    p("Visualize/hide one or more regions on the graph by clicking on the items in the legend. Double click to only visualize the selected item."),
+    p("Click on the autoscale button (the third one) to maximize the size of the graph."),
 
     h3("Time series"),
     fluidRow(
       box(title = "Cumulative events", width = 12,
-        mod_ts_prv_ui("ts_prv_cum")
+        eng_mod_ts_prv_ui("ts_prv_cum")
       )
     ),
 
     fluidRow(
       box(title = "New events", width = 12,
-        mod_ts_prv_ui("ts_prv_inc")
+        eng_mod_ts_prv_ui("ts_prv_inc")
       )
     )
   ),
@@ -397,7 +443,7 @@ eng_dashboard_body <- function() {dashboardBody(
 
     p("For the app development the following expansion packages were used: {shiny} ver. 1.4.0, {shinydashboard} v.0.7.1 and {golem} ver. 0.2.1."),
 
-    p("The analysis were performed using functions of the following packages: {stats} ver. 3.6.3, and {gam} ver. 1.16.1"),
+    p("The analyses were performed using functions of the following packages: {stats} ver. 3.6.3, and {gam} ver. 1.16.1"),
 
     p("Our graphs were produced with the following packages: {ggplot2} ver. 3.3.0 and {plotly} ver. 4.9.2."),
 
@@ -414,7 +460,7 @@ eng_dashboard_body <- function() {dashboardBody(
 
 
   tabItem(tabName = "impact",
-    h1("Pricipal indeces"),
+    h1("Principal indices"),
     eng_mod_ind_ita_ui("20200315")
   )