+++ R/eng-mod_ts_ita.R
@@ -10,17 +10,20 @@
 #' @importFrom plotly plotlyOutput
 eng_mod_ts_ita_ui <- function(id, title, width = 12){
   ns <- NS(id)
+  fluidPage(
+    fluidRow(shiny::checkboxInput(ns("y_log"), "Logarithmic scale")),
   fluidRow(
     box(title = title, width = width,
       plotlyOutput(ns("ts_plot"))
     )
   )
+ )
 }
 
 #' ts_ita Server Function
 #'
 #' @import ggplot2
-#' @importFrom plotly renderPlotly ggplotly
+#' @importFrom plotly renderPlotly ggplotly config
 #' @noRd
 eng_mod_ts_ita_server <- function(id, type = c("cum", "inc")) {
   type <- match.arg(type)
@@ -28,8 +31,7 @@ eng_mod_ts_ita_server <- function(id, type = c("cum", "inc")) {
   dpc_data <- dpc_covid19_ita_andamento_nazionale %>%
     dplyr::mutate(data = as.Date(.data$data))
 
-  var_to_exclude <- c("stato", "tamponi", "nuovi_attualmente_positivi")
-  var_of_interest <- setdiff(names(dpc_data), var_to_exclude)
+  var_of_interest <- c("data", measures("national"))
   exclude_from_pivoting <- "data"
 
   ts_data_to_plot <- dpc_data[var_of_interest] %>%
@@ -39,11 +41,10 @@ eng_mod_ts_ita_server <- function(id, type = c("cum", "inc")) {
     ) %>%
     dplyr::mutate(
       Measure = factor(.data$Measure,
-        levels = var_of_interest,
-        labels = var_of_interest %>%
-          stringr::str_replace_all("_", " ") %>%
-          stringr::str_to_title()
-      )
+        levels = measures("national"),
+        labels = measures("national") %>%
+          measure_to_labels(lang = "eng")
+    )
     )
 
   y_lab <- "N"
@@ -75,7 +76,20 @@ eng_mod_ts_ita_server <- function(id, type = c("cum", "inc")) {
 
   callModule(id = id, function(input, output, session) {
     ns <- session$ns
-    output$ts_plot <- renderPlotly(ggplotly(gg))
+    output$ts_plot <- renderPlotly({
+      if (input$y_log) {
+        gg <- gg + scale_y_continuous(
+          trans = 'log2',
+          breaks = scales::trans_breaks("log2", function(x) 2^x),
+          labels = scales::trans_format("log2", scales::math_format(2^.data[[".x"]]))
+        ) +
+          ylab(paste0(y_lab," - log2"))
+      }
+
+      ggplotly(gg) %>%
+        config(modeBarButtonsToRemove = c("zoomIn2d", "zoomOut2d", "pan2d", "select2d", "lasso2d")) %>%
+        config(displaylogo = FALSE)
+    })
   })
 }