+++ dev/02_dev.R
@@ -26,8 +26,6 @@ usethis::use_package("plotly")
 golem::add_module(name = "ts_ita")
 golem::add_module(name = "ts_reg")
 golem::add_module(name = "ts_prv")
-golem::add_module(name = "impact_veneto_a")
-golem::add_module(name = "impact_veneto_b")
 golem::add_module(name = "info_header")
 golem::add_module(name = "img_header")
 golem::add_module(name = "help_plot")
@@ -38,6 +36,9 @@ golem::add_module(name = "focus_20200318_piemonte")
 golem::add_module(name = "focus_20200318_veneto_intensive")
 golem::add_module(name = "focus_20200320_novara")
 golem::add_module(name = "focus_20200323_picco")
+golem::add_module(name = "focus_20200325")
+golem::add_module(name = "focus_20200327")
+golem::add_module(name = "focus_20200331")
 
 
 ## Add helper functions ----
@@ -48,7 +49,7 @@ golem::add_utils("render_box_using")
 golem::add_utils("extract_ci_from_gg_txt")
 golem::add_fct("cumulate_for_days")
 golem::add_utils("pred_to_tbl")
-
+golem::add_fct("regional_poiss")